requirements.txt
PyQt5>=5.15
yt-dlp>=2024.1
requests>=2.28
feedparser>=6.0.0,<7.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
